requirements.txt
unicornhat
unicornhatmini
flask
flask_cors
gpiozero
jsmin
RPi.GPIO
python-dotenv
requests
